ARM: qcom: automatically select PCI_DOMAINS if PCI is enabled
If multiple PCIe devices are present in the system, the kernel will
panic at boot time when trying to scan the PCI buses. This happens on
IPQ806x based platforms, which has 3 PCIe ports.

Enabling this option allows the kernel to assign the pci-domains
according to the device-tree content. This allows multiple PCIe
controllers to coexist in the system.

Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>