mtd: nand: raw: qcom_nandc: add boot_layout_mode support
ipq806x nand have a special ecc configuration for the boot pages. The
use of the non-boot pages configuration on boot pages cause I/O error
and can cause broken data written to the nand. Add support for this
special configuration if the page to be read/write is in the size of the
boot pages set by the dts.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>